barcode-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
piexif==1.1.3
exifread==3.0.0
zxing-cpp==2.2.1
easyocr==1.7.0
openai==1.3.6
rapidfuzz==3.5.2
requests==2.31.0
python-dotenv==1.0.0
